bug: Change data type for onset and peak_position from int to float

+++ cait/versatile/eventfunctions/scalarfunctions/mainparameters.py
@@ -69,8 +69,8 @@ class MainParameters(FncBaseClass):
     """
     params = (
             ("pulse_height", float),
-            ("peak_position", int),
-            ("onset", int),
+            ("peak_position", float),
+            ("onset", float),
             ("rise_time", float),
             ("decay_time", float),
             ("rms", float),
